requirements.txt
requests==2.25.1
urllib3==1.26.2
pytest==6.2.1
PyYAML==5.3.1
PyMySQL==0.10.1
pycryptodome==3.9.9
pytest-assume==2.4.2
Pillow==8.1.0
jsonpath==0.82
allure-pytest==2.8.29
pytest-html==3.1.1
pytest-xdist==2.2.0
pytest-ordering==0.6
pytest-rerunfailures==9.1.1
ddt==1.4.1
xlsxwriter==1.3.7
pandas==1.2.0
redis~=3.5.3
numpy~=1.19.5
Faker~=6.6.2
openpyxl==3.0.7
filelock==3.0.12
pdfplumber~=0.5.28
python-dateutil~=2.8.1